rename viewcontent to contentgrid
